+++ Code/AT_1Init.lua
@@ -778,6 +778,10 @@ function OnMsg.ClassesGenerate()
   	end -- if self.AT_enabled
   end  -- SupplyRocket:GenerateDepartures()
 
+<<<<<<< HEAD
+=======
+
+>>>>>>> develop
 end -- OnMsg.ClassesGenerate()
 
 -----------------------------------------------------------------------------------------------------------------------------------------