+++ metadata.lua
@@ -1,8 +1,14 @@
 return PlaceObj('ModDef', {
 	'title', "Automated Tourism",
+<<<<<<< HEAD
 	'description', "Automated Tourism\nv1.9.0\nhttps://steamcommunity.com/sharedfiles/filedetails/?id=1736068322\nby SKiRich",
 	'image', "AutomatedTourism.png",
 	'last_changes', "v1.9.0 March 27th, 2021",
+=======
+	'description', "Automated Tourism\nv1.8.4\nhttps://steamcommunity.com/sharedfiles/filedetails/?id=1736068322\nby SKiRich",
+	'image', "AutomatedTourism.png",
+	'last_changes', "v1.8.4 March 17th 2021",
+>>>>>>> develop
 	'ignore_files', {
 		"*.git/*",
 		"*.svn/*",
@@ -14,14 +20,24 @@ return PlaceObj('ModDef', {
 	'pops_desktop_uuid', "c515e2e3-ff0a-4dc2-91f4-c6b58431e755",
 	'pops_any_uuid', "44a35da1-f32e-4034-a8e5-4280e705a553",
 	'author', "SkiRich",
+<<<<<<< HEAD
 	'version', 77,
 	'lua_revision', 233360,
 	'saved_with_revision', 1001569,
+=======
+	'version', 74,
+	'lua_revision', 233360,
+	'saved_with_revision', 1001539,
+>>>>>>> develop
 	'code', {
 		"Code/AT_1Init.lua",
 		"Code/AT_2Panels.lua",
 		"Code/AT_3ModConfig.lua",
 	},
+<<<<<<< HEAD
 	'saved', 1616920900,
+=======
+	'saved', 1616006365,
+>>>>>>> develop
 	'screenshot1', "Images/AutomatedTourism.png",
 })